MM-18740 Sync app badge number when opening a push notification
